requirements.txt
torch>=2.0.0
transformers>=4.36.0
gradio>=4.0.0
PyPDF2>=3.0.0
groq>=0.31.0
accelerate>=0.25.0
sentencepiece>=0.1.99
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0